requirements.txt
appdirs==1.4.4
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
blessings==1.7
boto3==1.24.17
botocore==1.27.17
bpython==0.22.1
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
cloudinary==1.29.0
colorama==0.4.4
cryptography==37.0.4
curtsies==0.3.10
cwcwidth==0.1.6
decorator==5.1.1
deepdiff==5.8.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-crontab==0.7.1
django-filter==21.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-storages==1.12.3
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
executing==0.8.3
ffmpeg-python==0.2.0
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib==1.0.4
ipython==8.3.0
jedi==0.18.1
jmespath==1.0.1
jsonfield==3.1.0
jwt==1.3.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.22.4
openpyxl==3.0.10
ordered-set==4.1.0
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pydub==0.25.1
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pyxdg==0.27
pyxlsb==1.0.9
regex==2021.7.6
requests==2.27.1
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
SpeechRecognition==3.8.1
sqlparse==0.4.1
stack-data==0.2.0
tomli==1.0.4
traitlets==5.2.0
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
xlsx2csv==0.7.8
